requirements.txt
# BioFractal AI - Core Dependencies
# ==================================

# Core ML/AI Libraries
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
scipy>=1.10.0

# Neuroscience & Biology
brian2>=2.5.4
nest-desktop>=3.3.0
elephant>=0.13.0

# Quantum Computing
amazon-braket-sdk>=1.60.0
qiskit>=0.44.0
cirq>=1.2.0
pennylane>=0.32.0

# Scientific Computing
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
networkx>=3.1.0
scikit-learn>=1.3.0

# Data Processing
pandas>=2.0.0
h5py>=3.9.0
pymongo>=4.5.0

# Web & Visualization
streamlit>=1.25.0
dash>=2.12.0
bokeh>=3.2.0

# Async & Performance
asyncio-mqtt>=0.14.0
uvloop>=0.17.0
aiofiles>=23.0.0

# Configuration & Logging
pydantic>=2.0.0
python-dotenv>=1.0.0
loguru>=0.7.0
typer>=0.9.0

# Testing & Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0

# Optional Advanced Features
faiss-cpu>=1.7.4  # For similarity search
redis>=4.6.0      # For caching
celery>=5.3.0     # For distributed tasks